created misc_util
