+++ pydatastructs/__init__.py
@@ -1,3 +1,4 @@
 from .linear_data_structures import *
 from .trees import *
 from .miscellaneous_data_structures import *
+from .utils import *
