+++ pydatastructs/trees/binary_trees.py
@@ -1,42 +1,11 @@
 from __future__ import print_function, division
+from pydatastructs.utils import Node
 
 __all__ = [
     'Node',
     'BinaryTree',
     'BinarySearchTree'
 ]
-
-class Node(object):
-    """
-    Represents node in trees.
-
-    Parameters
-    ==========
-
-    data
-        Any valid data to be stored in the node.
-    key
-        Required for comparison operations.
-    left: int
-        Optional, index of the left child node.
-    right: int
-        Optional, index of the right child node.
-    """
-
-    __slots__ = ['key', 'data', 'left', 'right', 'is_root']
-
-    def __new__(cls, key, data):
-        obj = object.__new__(cls)
-        obj.data, obj.key = data, key
-        obj.left, obj.right = None, None
-        obj.is_root = False
-        return obj
-
-    def __str__(self):
-        """
-        Used for printing.
-        """
-        return str((self.left, self.key, self.data, self.right))
 
 class BinaryTree(object):
     """
