+++ pydatastructs/trees/space_partitioning_trees.py
@@ -1,4 +1,4 @@
-from pydatastructs.trees.binary_trees import Node
+from pydatastructs.utils import Node
 from collections import deque as Queue
 from pydatastructs.linear_data_structures.arrays import _check_type
