+++ pydatastructs/utils/__init__.py
@@ -1,0 +1,7 @@
+__all__ = []
+
+from . import misc_util
+from .misc_util import (
+    Node
+)
+__all__.extend(misc_util.__all__)
